+++ src/components/fleet/molecules/ShipsList/toggle-fleet/index.tsx
@@ -1,27 +1,12 @@
 import { ToggleButton, ToggleButtonGroup } from "@material-ui/core";
-import { makeStyles } from "@material-ui/styles";
 import { FC, MouseEvent } from "react";
 import { FleetNo } from "../../../../../models/ship";
-
-const useStyles = makeStyles(() => ({
-  root: {
-    "&.MuiToggleButtonGroup-root": {
-      display: "flex",
-    },
-
-    "& .MuiToggleButton-root": {
-      flexGrow: 1,
-    },
-  },
-}));
 
 type Props = {
   value: FleetNo;
   onChange: (fleetNo: FleetNo) => void;
 };
 export const ToggleFleet: FC<Props> = ({ value, onChange }) => {
-  const classes = useStyles();
-
   const handlerOnSelect = (_: MouseEvent, fleetNo: FleetNo) => {
     if (fleetNo !== null) {
       onChange(fleetNo);
@@ -32,10 +17,10 @@ export const ToggleFleet: FC<Props> = ({ value, onChange }) => {
     <ToggleButtonGroup
       value={value}
       exclusive
+      fullWidth
       size="small"
       color="primary"
       onChange={handlerOnSelect}
-      className={classes.root}
     >
       <ToggleButton value={0}>第一艦隊</ToggleButton>
       <ToggleButton value={1}>第二艦隊</ToggleButton>
