+++ src/components/home/create-new-fleet/index.tsx
@@ -1,31 +1,14 @@
 import { Fab, Grid } from "@material-ui/core";
 import { Add } from "@material-ui/icons";
-import { makeStyles } from "@material-ui/styles";
 import Link from "next/link";
 import { FC } from "react";
 
-const useStyles = makeStyles((theme) => ({
-  container: {
-    padding: theme.spacing(3),
-
-    ["@media (max-width: 719px)"]: {
-      padding: theme.spacing(2),
-    },
-  },
-
-  extendedIcon: {
-    marginRight: theme.spacing(1),
-  },
-}));
-
 export const CreateNewFleet: FC = () => {
-  const classes = useStyles();
-
   return (
-    <Grid container justifyContent="center" className={classes.container}>
+    <Grid container justifyContent="center" padding={{ xs: 2, sm: 3 }}>
       <Link href="/new" passHref>
         <Fab variant="extended" color="primary">
-          <Add className={classes.extendedIcon} />
+          <Add sx={{ mr: 1 }} />
           編成を作成
         </Fab>
       </Link>
